+++ examples/toggle_line_value/main.go
@@ -16,26 +16,27 @@ import (
 	"time"
 
 	"github.com/warthog618/go-gpiocdev"
-	"github.com/warthog618/go-gpiocdev/device/rpi"
 )
 
-// This example drives GPIO 22, which is pin J8-15 on a Raspberry Pi.
+// This example drives line 22 on gpiochip0.
 // The pin is toggled high and low at 1Hz with a 50% duty cycle.
-// Do not run this on a device which has this pin externally driven.
+// DO NOT run this on a device which has this pin externally driven.
 func main() {
-	offset := rpi.J8p15
+	offset := 22
+	chip := "gpiochip0"
 	v := 0
-	l, err := gpiocdev.RequestLine("gpiochip0", offset, gpiocdev.AsOutput(v))
+	l, err := gpiocdev.RequestLine(chip, offset, gpiocdev.AsOutput(v))
 	if err != nil {
 		panic(err)
 	}
 	// revert line to input on the way out.
 	defer func() {
 		l.Reconfigure(gpiocdev.AsInput)
+		fmt.Printf("Input pin %s:%d\n", chip, offset)
 		l.Close()
 	}()
 	values := map[int]string{0: "inactive", 1: "active"}
-	fmt.Printf("Set pin %d %s\n", offset, values[v])
+	fmt.Printf("Set pin %s:%d %s\n", chip, offset, values[v])
 
 	// capture exit signals to ensure pin is reverted to input on exit.
 	quit := make(chan os.Signal, 1)
@@ -44,10 +45,10 @@ func main() {
 
 	for {
 		select {
-		case <-time.After(2 * time.Second):
+		case <-time.After(500 * time.Millisecond):
 			v ^= 1
 			l.SetValue(v)
-			fmt.Printf("Set pin %d %s\n", offset, values[v])
+			fmt.Printf("Set pin %s:%d %s\n", chip, offset, values[v])
 		case <-quit:
 			return
 		}